# config/documentation/swagger-admin.json
{
	"swagger": "2.0",
	"info": {
		"title": "honorVet Project - API",
		"description": "honorVet Project -  Admin API Documentation",
		"version": "1.0"
	},
	"produces": [
		"application/json"
	],
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"paths": {
...
		  "/experience/addExperience": {
			"post": {
			  "summary": "User Work Experience add",
			  "tags": ["Work Experience"],
			  "content": ["application/json"],
			  "parameters": [
				{
				  "name": "userId",
				  "in": "formData",
				  "description": "Enter user ID",
				  "required": "true",
				  "type": "string"
				},
				{
				  "name": "jobTitle",
				  "in": "formData",
				  "description": "Enter job Title detail",
				  "required": false,
				  "type": "string"
				},
				{
				  "name": "company",
				  "in": "formData",
				  "description": "Enter company detials",
				  "required": "true",
				  "type": "string"
				},
				{
					"name": "location",
					"in": "formData",
					"description": "Enter your location",
					"required": "true",
					"type": "string"
				  },
				{
				  "name": "from",
				  "in": "formData",
				  "description": "Enter the date when you start",
				  "required": "true",
				  "type": "string"
				},
				{
				  "name": "to",
				  "in": "formData",
				  "description": "Enter the date when you end",
				  "required": "true",
				  "type": "string"
				},
				{
					"name": "employmentType",
					"in": "formData",
					"description": "Enter your employmentType",
					"enum":["Full_time","Part_time","Contractual","Seasonal"],
					"required": "true",
					"type": "string"
				  },
				  {
					"name": "jobDescription",
					"in": "formData",
					"description": "Enter your job Description",
					"required": "true",
					"type": "string"
				  }
			  ],
	  
			  "responses": {
				"200": {
				  "description": "Education add Successfull",
				  "schema": {
					"type": "object"
				  }
				},
				"405": {
				  "description": "Invalid input"
				}
			  }
			}
		  },
...
	    "/experience/editExperience": {
			"put": {
			  "summary": "User Work Experience edit",
			  "tags": ["Work Experience"],
			  "content": ["application/json"],
			  "parameters": [
				{
				  "name": "id",
				  "in": "formData",
				  "description": "Enter Work Experience ID",
				  "required": "true",
				  "type": "string"
				},
				{
					"name": "jobTitle",
					"in": "formData",
					"description": "Enter job Title detail",
					"required": false,
					"type": "string"
				  },
				  {
					"name": "company",
					"in": "formData",
					"description": "Enter company detials",
					"required": "true",
					"type": "string"
				  },
				  {
					"name": "location",
					"in": "formData",
					"description": "Enter your location",
					"required": "true",
					"type": "string"
				  },
				  {
					"name": "from",
					"in": "formData",
					"description": "Enter the date when you start",
					"required": "true",
					"type": "string"
				  },
				  {
					"name": "to",
					"in": "formData",
					"description": "Enter the date when you end",
					"required": "true",
					"type": "string"
				  },
				  {
					  "name": "employmentType",
					  "in": "formData",
					  "description": "Enter your employmentType",
					  "required": "true",
					  "type": "string"
					},
					{
					  "name": "jobDescription",
					  "in": "formData",
					  "description": "Enter your job Description",
					  "required": "true",
					  "type": "string"
					}
			  ],
	  
			  "responses": {
				"200": {
				  "description": "Education edit Successfull",
				  "schema": {
					"type": "object"
				  }
				},
				"405": {
				  "description": "Invalid input"
				}
			  }
			}
		  },
		 "/experience/deleteExperience": {
			"delete": {
			  "summary": "User Work Experience deleted",
			  "tags": ["Work Experience"],
			  "content": ["application/json"],
			  "parameters": [
				{
				  "name": "id",
				  "in": "formData",
				  "description": "Enter education ID",
				  "required": "true",
				  "type": "string"
				},
				{
					"name": "isDeleted",
					"in": "formData",
					"description": "Enter isDeleted value",
					"enum":[0,1],
					"required": "true",
					"type": "integer"
				  }
				
			  ],
	  
			  "responses": {
				"200": {
				  "description": "Education deleted Successfull",
				  "schema": {
					"type": "object"
				  }
				},
				"405": {
				  "description": "Invalid input"
				}
			  }
			}
		  } ,
...
			"/certificate/addcertification": {
				"post": {
				  "summary": "User Resume certification add",
				  "tags": ["Licenses and certification"],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "userId",
					  "in": "formData",
					  "description": "Enter user ID",
					  "required": "true",
					  "type": "string",
					  "value":"80d874f6-c488-4b1c-bb1a-946297a542a2"
					},
					{
					  "name": "name",
					  "in": "formData",
					  "description": "Enter your name",
					  "required": false,
					  "type": "string",
					  "value":"vaneet kumar"
					},
					{
					  "name": "attestedby",
					  "in": "formData",
					  "description": "Enter who attested the certificate",
					  "required": "true",
					  "type": "string",
					  "value":"Government of India"
					},
					{
					  "name": "issuedate",
					  "in": "formData",
					  "description": "Enter issuedate ",
					  "required": "true",
					  "type": "string",
					  "value":"2019-09-09"
					},
					{
					  "name": "expirydate",
					  "in": "formData",
					  "description": "Enter expiry date",
					  "required": "true",
					  "type": "string",
					  "value":"2030-09-09"
					},
					{
						"name": "certificationo",
						"in": "formData",
						"description": "Enter certificatio no",
						"required": "true",
						"type": "string",
						"value":"AEXP6789"
					  },
					  {
						"name": "credentialURL",
						"in": "formData",
						"description": "Enter credential URL",
						"required": "true",
						"type": "string",
						"value":"http://locahost:3002"
					  }
				  ],
		  
				  "responses": {
					"200": {
					  "description": "Certificate add Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			"/certificate/getcertification": {
				"get": {
					"x-swagger-router-controller": "middleware-name1",
					"summary": "User Resume certificate get all details",
					"tags": [
					  "Licenses and certification"
					],
					"description": "Get Volunteer Experience List",
					"security": [
						{
							"Bearer": []
						}
					],
					"responses": {}
				}
			},
			"/certificate/editcertification": {
				"put": {
				  "summary": "User Resume certification edit",
				  "tags": ["Licenses and certification"],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "id",
					  "in": "formData",
					  "description": "Enter Award and Honour ID",
					  "required": "true",
					  "type": "string",
					  "value":"53becdf3-c284-4fdc-9239-98104a97ba1a"
					},
					{
						"name": "name",
						"in": "formData",
						"description": "Enter your name",
						"required": false,
						"type": "string",
						"value":"Vaneet thakur"
					  },
					  {
						"name": "attestedby",
						"in": "formData",
						"description": "Enter who attested the certificate",
						"required": "true",
						"type": "string",
						"value":"Govenment of HP"
					  },
					  {
						"name": "issuedate",
						"in": "formData",
						"description": "Enter issuedate ",
						"required": "true",
						"type": "string",
						"value":"2015-09-09"
					  },
					  {
						"name": "expirydate",
						"in": "formData",
						"description": "Enter expiry date",
						"required": "true",
						"type": "string",
						"value":"2040-09-09"
					  },
					  {
						  "name": "certificationo",
						  "in": "formData",
						  "description": "Enter certificatio no",
						  "required": "true",
						  "type": "string",
						  "value":"EXPRP6789"
						},
						{
						  "name": "credentialURL",
						  "in": "formData",
						  "description": "Enter credential URL",
						  "required": "true",
						  "type": "string",
						  "value":"www.google.com"
						}
				  ],
		  
				  "responses": {
					"200": {
					  "description": "Certificate edit Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			 "/certificate/deletecertification": {
				"delete": {
				  "summary": "User Resume certification deleted",
				  "tags": ["Licenses and certification"],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "id",
					  "in": "formData",
					  "description": "Enter volunteer Experience ID",
					  "required": "true",
					  "type": "string",
					  "value":"53becdf3-c284-4fdc-9239-98104a97ba1a"
					},
					{
						"name": "isDeleted",
						"in": "formData",
						"description": "Enter isDeleted value",
						"enum":[0,1],
						"required": "true",
						"type": "integer"
					  }
					
				  ],
		  
				  "responses": {
					"200": {
					  "description": "Certificate deleted Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			  "/language/addlanguage": {
				"post": {
				  "summary": "User Register",
				  "tags": ["User Language"],
				  "content": ["application/json"],
				  "parameters": [
					{
						"name": "userId",
						"in": "formData",
						"description": "Enter user ID",
						"required": "true",
						"type": "string",
						"value":"80d874f6-c488-4b1c-bb1a-946297a542a2"
					  },
					{
					  "name": "language",
					  "in": "formData",
					  "description": "Enter user language",
					  "required": "true",
					  "type": "string",
					  "value":"English"
					},
					{
					  "name": "proficiency",
					  "in": "formData",
					  "description": "Enter your proficiency",
					  "required": false,
					  "type": "string",
					  "enum":["Beginner","Proficient","Expert"],
					  "value":""
					}
				  ],
		  
				  "responses": {
					"200": {
					  "description": "language add Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			  "/language/getlanguage": {
				"get": {
					"x-swagger-router-controller": "middleware-name1",
					"summary": "User language get all details",
					"tags": [
					  "User Language"
					],
					"description": "Get Volunteer Experience List",
					"security": [
						{
							"Bearer": []
						}
					],
					"responses": {}
				}
			},
			"/language/editlanguage": {
				"put": {
				  "summary": "User Resume language edit",
				  "tags": ["User Language"],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "id",
					  "in": "formData",
					  "description": "Enter Award and Honour ID",
					  "required": "true",
					  "type": "string",
					  "value":"377ef70e-3918-4e0f-a895-b456f81aa758"
					},
					{
						"name": "language",
						"in": "formData",
						"description": "Enter user language",
						"required": "true",
						"type": "string",
						"value":"English"
					  },
					  {
						"name": "proficiency",
						"in": "formData",
						"description": "Enter your proficiency",
						"required": false,
						"type": "string",
						"enum":["Beginner","Proficient","Expert"],
						"value":""
					  }
				  ],
		  
				  "responses": {
					"200": {
					  "description": "language edit Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			 "/language/deletelanguage": {
				"delete": {
				  "summary": "User Resume langugae deleted",
				  "tags": ["User Language"],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "id",
					  "in": "formData",
					  "description": "Enter volunteer Experience ID",
					  "required": "true",
					  "type": "string",
					  "value":"377ef70e-3918-4e0f-a895-b456f81aa758"
					},
					{
						"name": "isDeleted",
						"in": "formData",
						"description": "Enter isDeleted value",
						"enum":[0,1],
						"required": "true",
						"type": "integer"
					  }
					
				  ],
		  
				  "responses": {
					"200": {
					  "description": "language deleted Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			  "/projectUnderTaken/addProjectUnderTaken": {
				"post": {
				  "summary": "User Resume Project under taken add",
				  "tags": ["Project under taken "],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "userId",
					  "in": "formData",
					  "description": "Enter user ID",
					  "required": "true",
					  "type": "string",
					  "value":"80d874f6-c488-4b1c-bb1a-946297a542a2"
					},
					{
					  "name": "associatedWith",
					  "in": "formData",
					  "description": "Enter who is associated with you with your project detail",
					  "required": false,
					  "type": "string",
					  "value":"Sourabh kumar"
					},
					{
					  "name": "projectTitle",
					  "in": "formData",
					  "description": "Enter your project Title detials",
					  "required": "true",
					  "type": "string",
					  "value":"Honovet"
					},
					{
					  "name": "from",
					  "in": "formData",
					  "description": "Enter the date when you start",
					  "required": "true",
					  "type": "string",
					  "value":"2022-12-06"
					},
					{
					  "name": "to",
					  "in": "formData",
					  "description": "Enter the date when you end",
					  "required": "true",
					  "type": "string",
					  "value":"2022-12-08"
					},
					{
						"name": "location",
						"in": "formData",
						"description": "Enter your location",
						"required": "true",
						"type": "string",
						"value":"Applify Mohali"
					  },
					  {
						  "name": "currentlyOnGoing",
						  "in": "formData",
						  "description": "Enter your currently On Going status",
						  "required": "true",
						  "enum":[0,1],
						  "type": "string"
						},
						{
							"name": "projectDetails",
							"in": "formData",
							"description": "Enter your project Details",
							"required": "true",
							"type": "string",
							"value":"It is job finder project"
						  }
				  ],
		  
				  "responses": {
					"200": {
					  "description": "Project under taken add Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			"/projectUnderTaken/getProjectUnderTaken": {
				"get": {
					"x-swagger-router-controller": "middleware-name1",
					"summary": "User Resume project under taken get all details",
					"tags": [
						"Project under taken "
					],
					"description": "Get Project under taken List",
					"security": [
						{
							"Bearer": []
						}
					],
					"responses": {}
				}
			},
			"/projectUnderTaken/editProjectUnderTaken": {
				"put": {
				  "summary": "User Resume Project under taken edit",
				  "tags": ["Project under taken "],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "id",
					  "in": "formData",
					  "description": "Enter Volunteer Experience ID",
					  "required": "true",
					  "type": "string",
					  "value":"e71fbbae-2f5c-43a2-9d7c-981139c12c39"
					},
					{
						"name": "associatedWith",
						"in": "formData",
						"description": "Enter who is associated with you with your project detail",
						"required": false,
						"type": "string",
						"value":"Self"
					  },
					  {
						"name": "projectTitle",
						"in": "formData",
						"description": "Enter your project Title detials",
						"required": "true",
						"type": "string",
						"value":"Honovet IOS APP"
					  },
					  {
						"name": "from",
						"in": "formData",
						"description": "Enter the date when you start",
						"required": "true",
						"type": "string",
						"value":"2022-12-06"
					  },
					  {
						"name": "to",
						"in": "formData",
						"description": "Enter the date when you end",
						"required": "true",
						"type": "string",
						"value":"2022-12-08"
					  },
					  {
						  "name": "location",
						  "in": "formData",
						  "description": "Enter your location",
						  "required": "true",
						  "type": "string",
						  "value":"Applify"
						},
						{
							"name": "currentlyOnGoing",
							"in": "formData",
							"description": "Enter your currently On Going status",
							"required": "true",
							"enum":[0,1],
							"type": "string"
						  },
						  {
							  "name": "projectDetails",
							  "in": "formData",
							  "description": "Enter your project Details",
							  "required": "true",
							  "type": "string",
							  "value":"Create for self practice"
							}
				  ],
		  
				  "responses": {
					"200": {
					  "description": "Project under taken edit Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  },
			 "/projectUnderTaken/deleteProjectUnderTaken": {
				"delete": {
				  "summary": "User Resume Project under taken deleted",
				  "tags": ["Project under taken "],
				  "content": ["application/json"],
				  "parameters": [
					{
					  "name": "id",
					  "in": "formData",
					  "description": "Enter Project Under taken ID",
					  "required": "true",
					  "type": "string",
					  "value":"e71fbbae-2f5c-43a2-9d7c-981139c12c39"
					},
					{
						"name": "isDeleted",
						"in": "formData",
						"description": "Enter isDeleted value",
						"enum":[0,1],
						"required": "true",
						"type": "integer"
					  }
					
				  ],
		  
				  "responses": {
					"200": {
					  "description": "Project under taken details deleted Successfull",
					  "schema": {
						"type": "object"
					  }
					},
					"405": {
					  "description": "Invalid input"
					}
				  }
				}
			  }
	}, 
...
}